+++ lib/src/widgets/panels/overlay_panel.dart
@@ -35,6 +35,7 @@ class OverlayPanel extends BasePanel {
     super.iconColor,
     super.panelBoxDecoration,
     super.headerDecoration,
+    super.clipContent = false,
     super.key,
   }) : super(initialCollapsed: false);
 
@@ -80,7 +81,10 @@ class OverlayPanel extends BasePanel {
     return Column(
       crossAxisAlignment: CrossAxisAlignment.stretch,
       mainAxisSize: MainAxisSize.min,
-      children: [if (header != null) header, content],
+      children: [
+        if (header != null) header,
+        clipContent ? ClipRect(child: content) : content,
+      ],
     );
   }
 }